+++ sdk/managementgroups/arm-managementgroups/src/operations/managementGroups.ts
@@ -13,8 +13,12 @@ import * as coreClient from "@azure/core-client";
 import * as Mappers from "../models/mappers.js";
 import * as Parameters from "../models/parameters.js";
 import { ManagementGroupsAPI } from "../managementGroupsAPI.js";
-import { PollerLike, PollOperationState, LroEngine } from "@azure/core-lro";
-import { LroImpl } from "../lroImpl.js";
+import {
+  SimplePollerLike,
+  OperationState,
+  createHttpPoller,
+} from "@azure/core-lro";
+import { createLroSpec } from "../lroImpl.js";
 import {
   ManagementGroupInfo,
   ManagementGroupsListNextOptionalParams,
@@ -35,7 +39,7 @@ import {
   ManagementGroupsDeleteOptionalParams,
   ManagementGroupsDeleteResponse,
   ManagementGroupsListNextResponse,
-  ManagementGroupsGetDescendantsNextResponse
+  ManagementGroupsGetDescendantsNextResponse,
 } from "../models/index.js";
 
 /// <reference lib="esnext.asynciterable" />
@@ -57,7 +61,7 @@ export class ManagementGroupsImpl implements ManagementGroups {
    * @param options The options parameters.
    */
   public list(
-    options?: ManagementGroupsListOptionalParams
+    options?: ManagementGroupsListOptionalParams,
   ): PagedAsyncIterableIterator<ManagementGroupInfo> {
     const iter = this.listPagingAll(options);
     return {
@@ -72,13 +76,13 @@ export class ManagementGroupsImpl implements ManagementGroups {
           throw new Error("maxPageSize is not supported by this operation.");
         }
         return this.listPagingPage(options, settings);
-      }
+      },
     };
   }
 
   private async *listPagingPage(
     options?: ManagementGroupsListOptionalParams,
-    settings?: PageSettings
+    settings?: PageSettings,
   ): AsyncIterableIterator<ManagementGroupInfo[]> {
     let result: ManagementGroupsListResponse;
     let continuationToken = settings?.continuationToken;
@@ -99,7 +103,7 @@ export class ManagementGroupsImpl implements ManagementGroups {
   }
 
   private async *listPagingAll(
-    options?: ManagementGroupsListOptionalParams
+    options?: ManagementGroupsListOptionalParams,
   ): AsyncIterableIterator<ManagementGroupInfo> {
     for await (const page of this.listPagingPage(options)) {
       yield* page;
@@ -114,7 +118,7 @@ export class ManagementGroupsImpl implements ManagementGroups {
    */
   public listDescendants(
     groupId: string,
-    options?: ManagementGroupsGetDescendantsOptionalParams
+    options?: ManagementGroupsGetDescendantsOptionalParams,
   ): PagedAsyncIterableIterator<DescendantInfo> {
     const iter = this.getDescendantsPagingAll(groupId, options);
     return {
@@ -129,14 +133,14 @@ export class ManagementGroupsImpl implements ManagementGroups {
           throw new Error("maxPageSize is not supported by this operation.");
         }
         return this.getDescendantsPagingPage(groupId, options, settings);
-      }
+      },
     };
   }
 
   private async *getDescendantsPagingPage(
     groupId: string,
     options?: ManagementGroupsGetDescendantsOptionalParams,
-    settings?: PageSettings
+    settings?: PageSettings,
   ): AsyncIterableIterator<DescendantInfo[]> {
     let result: ManagementGroupsGetDescendantsResponse;
     let continuationToken = settings?.continuationToken;
@@ -151,7 +155,7 @@ export class ManagementGroupsImpl implements ManagementGroups {
       result = await this._getDescendantsNext(
         groupId,
         continuationToken,
-        options
+        options,
       );
       continuationToken = result.nextLink;
       let page = result.value || [];
@@ -162,7 +166,7 @@ export class ManagementGroupsImpl implements ManagementGroups {
 
   private async *getDescendantsPagingAll(
     groupId: string,
-    options?: ManagementGroupsGetDescendantsOptionalParams
+    options?: ManagementGroupsGetDescendantsOptionalParams,
   ): AsyncIterableIterator<DescendantInfo> {
     for await (const page of this.getDescendantsPagingPage(groupId, options)) {
       yield* page;
@@ -175,7 +179,7 @@ export class ManagementGroupsImpl implements ManagementGroups {
    * @param options The options parameters.
    */
   private _list(
-    options?: ManagementGroupsListOptionalParams
+    options?: ManagementGroupsListOptionalParams,
   ): Promise<ManagementGroupsListResponse> {
     return this.client.sendOperationRequest({ options }, listOperationSpec);
   }
@@ -188,11 +192,11 @@ export class ManagementGroupsImpl implements ManagementGroups {
    */
   get(
     groupId: string,
-    options?: ManagementGroupsGetOptionalParams
+    options?: ManagementGroupsGetOptionalParams,
   ): Promise<ManagementGroupsGetResponse> {
     return this.client.sendOperationRequest(
       { groupId, options },
-      getOperationSpec
+      getOperationSpec,
     );
   }
 
@@ -208,30 +212,29 @@ export class ManagementGroupsImpl implements ManagementGroups {
   async beginCreateOrUpdate(
     groupId: string,
     createManagementGroupRequest: CreateManagementGroupRequest,
-    options?: ManagementGroupsCreateOrUpdateOptionalParams
+    options?: ManagementGroupsCreateOrUpdateOptionalParams,
   ): Promise<
-    PollerLike<
-      PollOperationState<ManagementGroupsCreateOrUpdateResponse>,
+    SimplePollerLike<
+      OperationState<ManagementGroupsCreateOrUpdateResponse>,
       ManagementGroupsCreateOrUpdateResponse
     >
   > {
     const directSendOperation = async (
       args: coreClient.OperationArguments,
-      spec: coreClient.OperationSpec
+      spec: coreClient.OperationSpec,
     ): Promise<ManagementGroupsCreateOrUpdateResponse> => {
       return this.client.sendOperationRequest(args, spec);
     };
-    const sendOperation = async (
+    const sendOperationFn = async (
       args: coreClient.OperationArguments,
-      spec: coreClient.OperationSpec
+      spec: coreClient.OperationSpec,
     ) => {
-      let currentRawResponse:
-        | coreClient.FullOperationResponse
-        | undefined = undefined;
+      let currentRawResponse: coreClient.FullOperationResponse | undefined =
+        undefined;
       const providedCallback = args.options?.onResponse;
       const callback: coreClient.RawResponseCallback = (
         rawResponse: coreClient.FullOperationResponse,
-        flatResponse: unknown
+        flatResponse: unknown,
       ) => {
         currentRawResponse = rawResponse;
         providedCallback?.(rawResponse, flatResponse);
@@ -240,8 +243,8 @@ export class ManagementGroupsImpl implements ManagementGroups {
         ...args,
         options: {
           ...args.options,
-          onResponse: callback
-        }
+          onResponse: callback,
+        },
       };
       const flatResponse = await directSendOperation(updatedArgs, spec);
       return {
@@ -249,20 +252,23 @@ export class ManagementGroupsImpl implements ManagementGroups {
         rawResponse: {
           statusCode: currentRawResponse!.status,
           body: currentRawResponse!.parsedBody,
-          headers: currentRawResponse!.headers.toJSON()
-        }
+          headers: currentRawResponse!.headers.toJSON(),
+        },
       };
     };
 
-    const lro = new LroImpl(
-      sendOperation,
-      { groupId, createManagementGroupRequest, options },
-      createOrUpdateOperationSpec
-    );
-    const poller = new LroEngine(lro, {
-      resumeFrom: options?.resumeFrom,
+    const lro = createLroSpec({
+      sendOperationFn,
+      args: { groupId, createManagementGroupRequest, options },
+      spec: createOrUpdateOperationSpec,
+    });
+    const poller = await createHttpPoller<
+      ManagementGroupsCreateOrUpdateResponse,
+      OperationState<ManagementGroupsCreateOrUpdateResponse>
+    >(lro, {
+      restoreFrom: options?.resumeFrom,
       intervalInMs: options?.updateIntervalInMs,
-      lroResourceLocationConfig: "azure-async-operation"
+      resourceLocationConfig: "azure-async-operation",
     });
     await poller.poll();
     return poller;
@@ -280,12 +286,12 @@ export class ManagementGroupsImpl implements ManagementGroups {
   async beginCreateOrUpdateAndWait(
     groupId: string,
     createManagementGroupRequest: CreateManagementGroupRequest,
-    options?: ManagementGroupsCreateOrUpdateOptionalParams
+    options?: ManagementGroupsCreateOrUpdateOptionalParams,
   ): Promise<ManagementGroupsCreateOrUpdateResponse> {
     const poller = await this.beginCreateOrUpdate(
       groupId,
       createManagementGroupRequest,
-      options
+      options,
     );
     return poller.pollUntilDone();
   }
@@ -300,11 +306,11 @@ export class ManagementGroupsImpl implements ManagementGroups {
   update(
     groupId: string,
     patchGroupRequest: PatchManagementGroupRequest,
-    options?: ManagementGroupsUpdateOptionalParams
+    options?: ManagementGroupsUpdateOptionalParams,
   ): Promise<ManagementGroupsUpdateResponse> {
     return this.client.sendOperationRequest(
       { groupId, patchGroupRequest, options },
-      updateOperationSpec
+      updateOperationSpec,
     );
   }
 
@@ -317,30 +323,29 @@ export class ManagementGroupsImpl implements ManagementGroups {
    */
   async beginDelete(
     groupId: string,
-    options?: ManagementGroupsDeleteOptionalParams
+    options?: ManagementGroupsDeleteOptionalParams,
   ): Promise<
-    PollerLike<
-      PollOperationState<ManagementGroupsDeleteResponse>,
+    SimplePollerLike<
+      OperationState<ManagementGroupsDeleteResponse>,
       ManagementGroupsDeleteResponse
     >
   > {
     const directSendOperation = async (
       args: coreClient.OperationArguments,
-      spec: coreClient.OperationSpec
+      spec: coreClient.OperationSpec,
     ): Promise<ManagementGroupsDeleteResponse> => {
       return this.client.sendOperationRequest(args, spec);
     };
-    const sendOperation = async (
+    const sendOperationFn = async (
       args: coreClient.OperationArguments,
-      spec: coreClient.OperationSpec
+      spec: coreClient.OperationSpec,
     ) => {
-      let currentRawResponse:
-        | coreClient.FullOperationResponse
-        | undefined = undefined;
+      let currentRawResponse: coreClient.FullOperationResponse | undefined =
+        undefined;
       const providedCallback = args.options?.onResponse;
       const callback: coreClient.RawResponseCallback = (
         rawResponse: coreClient.FullOperationResponse,
-        flatResponse: unknown
+        flatResponse: unknown,
       ) => {
         currentRawResponse = rawResponse;
         providedCallback?.(rawResponse, flatResponse);
@@ -349,8 +354,8 @@ export class ManagementGroupsImpl implements ManagementGroups {
         ...args,
         options: {
           ...args.options,
-          onResponse: callback
-        }
+          onResponse: callback,
+        },
       };
       const flatResponse = await directSendOperation(updatedArgs, spec);
       return {
@@ -358,20 +363,23 @@ export class ManagementGroupsImpl implements ManagementGroups {
         rawResponse: {
           statusCode: currentRawResponse!.status,
           body: currentRawResponse!.parsedBody,
-          headers: currentRawResponse!.headers.toJSON()
-        }
+          headers: currentRawResponse!.headers.toJSON(),
+        },
       };
     };
 
-    const lro = new LroImpl(
-      sendOperation,
-      { groupId, options },
-      deleteOperationSpec
-    );
-    const poller = new LroEngine(lro, {
-      resumeFrom: options?.resumeFrom,
+    const lro = createLroSpec({
+      sendOperationFn,
+      args: { groupId, options },
+      spec: deleteOperationSpec,
+    });
+    const poller = await createHttpPoller<
+      ManagementGroupsDeleteResponse,
+      OperationState<ManagementGroupsDeleteResponse>
+    >(lro, {
+      restoreFrom: options?.resumeFrom,
       intervalInMs: options?.updateIntervalInMs,
-      lroResourceLocationConfig: "azure-async-operation"
+      resourceLocationConfig: "azure-async-operation",
     });
     await poller.poll();
     return poller;
@@ -386,7 +394,7 @@ export class ManagementGroupsImpl implements ManagementGroups {
    */
   async beginDeleteAndWait(
     groupId: string,
-    options?: ManagementGroupsDeleteOptionalParams
+    options?: ManagementGroupsDeleteOptionalParams,
   ): Promise<ManagementGroupsDeleteResponse> {
     const poller = await this.beginDelete(groupId, options);
     return poller.pollUntilDone();
@@ -400,11 +408,11 @@ export class ManagementGroupsImpl implements ManagementGroups {
    */
   private _getDescendants(
     groupId: string,
-    options?: ManagementGroupsGetDescendantsOptionalParams
+    options?: ManagementGroupsGetDescendantsOptionalParams,
   ): Promise<ManagementGroupsGetDescendantsResponse> {
     return this.client.sendOperationRequest(
       { groupId, options },
-      getDescendantsOperationSpec
+      getDescendantsOperationSpec,
     );
   }
 
@@ -415,11 +423,11 @@ export class ManagementGroupsImpl implements ManagementGroups {
    */
   private _listNext(
     nextLink: string,
-    options?: ManagementGroupsListNextOptionalParams
+    options?: ManagementGroupsListNextOptionalParams,
   ): Promise<ManagementGroupsListNextResponse> {
     return this.client.sendOperationRequest(
       { nextLink, options },
-      listNextOperationSpec
+      listNextOperationSpec,
     );
   }
 
@@ -432,11 +440,11 @@ export class ManagementGroupsImpl implements ManagementGroups {
   private _getDescendantsNext(
     groupId: string,
     nextLink: string,
-    options?: ManagementGroupsGetDescendantsNextOptionalParams
+    options?: ManagementGroupsGetDescendantsNextOptionalParams,
   ): Promise<ManagementGroupsGetDescendantsNextResponse> {
     return this.client.sendOperationRequest(
       { groupId, nextLink, options },
-      getDescendantsNextOperationSpec
+      getDescendantsNextOperationSpec,
     );
   }
 }
@@ -448,57 +456,57 @@ const listOperationSpec: coreClient.OperationSpec = {
   httpMethod: "GET",
   responses: {
     200: {
-      bodyMapper: Mappers.ManagementGroupListResult
+      bodyMapper: Mappers.ManagementGroupListResult,
     },
     default: {
-      bodyMapper: Mappers.ErrorResponse
-    }
+      bodyMapper: Mappers.ErrorResponse,
+    },
   },
   queryParameters: [Parameters.apiVersion, Parameters.skiptoken],
   urlParameters: [Parameters.$host],
   headerParameters: [Parameters.accept, Parameters.cacheControl],
-  serializer
+  serializer,
 };
 const getOperationSpec: coreClient.OperationSpec = {
   path: "/providers/Microsoft.Management/managementGroups/{groupId}",
   httpMethod: "GET",
   responses: {
     200: {
-      bodyMapper: Mappers.ManagementGroup
+      bodyMapper: Mappers.ManagementGroup,
     },
     default: {
-      bodyMapper: Mappers.ErrorResponse
-    }
+      bodyMapper: Mappers.ErrorResponse,
+    },
   },
   queryParameters: [
     Parameters.apiVersion,
     Parameters.expand,
     Parameters.recurse,
-    Parameters.filter
+    Parameters.filter,
   ],
   urlParameters: [Parameters.$host, Parameters.groupId],
   headerParameters: [Parameters.accept, Parameters.cacheControl],
-  serializer
+  serializer,
 };
 const createOrUpdateOperationSpec: coreClient.OperationSpec = {
   path: "/providers/Microsoft.Management/managementGroups/{groupId}",
   httpMethod: "PUT",
   responses: {
     200: {
-      bodyMapper: Mappers.ManagementGroup
+      bodyMapper: Mappers.ManagementGroup,
     },
     201: {
-      bodyMapper: Mappers.ManagementGroup
+      bodyMapper: Mappers.ManagementGroup,
     },
     202: {
-      bodyMapper: Mappers.ManagementGroup
+      bodyMapper: Mappers.ManagementGroup,
     },
     204: {
-      bodyMapper: Mappers.ManagementGroup
+      bodyMapper: Mappers.ManagementGroup,
     },
     default: {
-      bodyMapper: Mappers.ErrorResponse
-    }
+      bodyMapper: Mappers.ErrorResponse,
+    },
   },
   requestBody: Parameters.createManagementGroupRequest,
   queryParameters: [Parameters.apiVersion],
@@ -506,21 +514,21 @@ const createOrUpdateOperationSpec: coreClient.OperationSpec = {
   headerParameters: [
     Parameters.accept,
     Parameters.cacheControl,
-    Parameters.contentType
+    Parameters.contentType,
   ],
   mediaType: "json",
-  serializer
+  serializer,
 };
 const updateOperationSpec: coreClient.OperationSpec = {
   path: "/providers/Microsoft.Management/managementGroups/{groupId}",
   httpMethod: "PATCH",
   responses: {
     200: {
-      bodyMapper: Mappers.ManagementGroup
+      bodyMapper: Mappers.ManagementGroup,
     },
     default: {
-      bodyMapper: Mappers.ErrorResponse
-    }
+      bodyMapper: Mappers.ErrorResponse,
+    },
   },
   requestBody: Parameters.patchGroupRequest,
   queryParameters: [Parameters.apiVersion],
@@ -528,10 +536,10 @@ const updateOperationSpec: coreClient.OperationSpec = {
   headerParameters: [
     Parameters.accept,
     Parameters.cacheControl,
-    Parameters.contentType
+    Parameters.contentType,
   ],
   mediaType: "json",
-  serializer
+  serializer,
 };
 const deleteOperationSpec: coreClient.OperationSpec = {
   path: "/providers/Microsoft.Management/managementGroups/{groupId}",
@@ -539,83 +547,76 @@ const deleteOperationSpec: coreClient.OperationSpec = {
   responses: {
     200: {
       bodyMapper: Mappers.AzureAsyncOperationResults,
-      headersMapper: Mappers.ManagementGroupsDeleteHeaders
+      headersMapper: Mappers.ManagementGroupsDeleteHeaders,
     },
     201: {
       bodyMapper: Mappers.AzureAsyncOperationResults,
-      headersMapper: Mappers.ManagementGroupsDeleteHeaders
+      headersMapper: Mappers.ManagementGroupsDeleteHeaders,
     },
     202: {
       bodyMapper: Mappers.AzureAsyncOperationResults,
-      headersMapper: Mappers.ManagementGroupsDeleteHeaders
+      headersMapper: Mappers.ManagementGroupsDeleteHeaders,
     },
     204: {
       bodyMapper: Mappers.AzureAsyncOperationResults,
-      headersMapper: Mappers.ManagementGroupsDeleteHeaders
+      headersMapper: Mappers.ManagementGroupsDeleteHeaders,
     },
     default: {
-      bodyMapper: Mappers.ErrorResponse
-    }
+      bodyMapper: Mappers.ErrorResponse,
+    },
   },
   queryParameters: [Parameters.apiVersion],
   urlParameters: [Parameters.$host, Parameters.groupId],
   headerParameters: [Parameters.accept, Parameters.cacheControl],
-  serializer
+  serializer,
 };
 const getDescendantsOperationSpec: coreClient.OperationSpec = {
-  path:
-    "/providers/Microsoft.Management/managementGroups/{groupId}/descendants",
+  path: "/providers/Microsoft.Management/managementGroups/{groupId}/descendants",
   httpMethod: "GET",
   responses: {
     200: {
-      bodyMapper: Mappers.DescendantListResult
+      bodyMapper: Mappers.DescendantListResult,
     },
     default: {
-      bodyMapper: Mappers.ErrorResponse
-    }
+      bodyMapper: Mappers.ErrorResponse,
+    },
   },
   queryParameters: [
     Parameters.apiVersion,
     Parameters.skiptoken,
-    Parameters.top
+    Parameters.top,
   ],
   urlParameters: [Parameters.$host, Parameters.groupId],
   headerParameters: [Parameters.accept],
-  serializer
+  serializer,
 };
 const listNextOperationSpec: coreClient.OperationSpec = {
   path: "{nextLink}",
   httpMethod: "GET",
   responses: {
     200: {
-      bodyMapper: Mappers.ManagementGroupListResult
+      bodyMapper: Mappers.ManagementGroupListResult,
     },
     default: {
-      bodyMapper: Mappers.ErrorResponse
-    }
+      bodyMapper: Mappers.ErrorResponse,
+    },
   },
-  queryParameters: [Parameters.apiVersion, Parameters.skiptoken],
   urlParameters: [Parameters.$host, Parameters.nextLink],
   headerParameters: [Parameters.accept, Parameters.cacheControl],
-  serializer
+  serializer,
 };
 const getDescendantsNextOperationSpec: coreClient.OperationSpec = {
   path: "{nextLink}",
   httpMethod: "GET",
   responses: {
     200: {
-      bodyMapper: Mappers.DescendantListResult
+      bodyMapper: Mappers.DescendantListResult,
     },
     default: {
-      bodyMapper: Mappers.ErrorResponse
-    }
+      bodyMapper: Mappers.ErrorResponse,
+    },
   },
-  queryParameters: [
-    Parameters.apiVersion,
-    Parameters.skiptoken,
-    Parameters.top
-  ],
   urlParameters: [Parameters.$host, Parameters.groupId, Parameters.nextLink],
   headerParameters: [Parameters.accept],
-  serializer
+  serializer,
 };